Added `RemoteDataItemMapper`
