+++ InstagramFeedCloneApp/API/Post.swift
@@ -1,0 +1,14 @@
+//
+//  Post.swift
+//  InstagramFeedCloneApp
+//
+//  Created by Mohamed Afsal on 17/04/2025.
+//
+
+import Foundation
+
+struct Post: Codable, Identifiable {
+    let id: String
+    let images: [Image]?
+}
+
